+++ packages/gasket-cli/test/unit/scaffold/actions/write-pkg.test.js
@@ -24,7 +24,11 @@ describe('write-pkg', () => {
     writeStub = sandbox.stub();
 
     writePkg = proxyquire('../../../../src/scaffold/actions/write-pkg', {
-      'util': { promisify: () => writeStub },
+      'fs': {
+        promises: {
+          writeFile: writeStub
+        }
+      },
       '../action-wrapper': require('../../../helpers').mockActionWrapper
     });
   });