+++ packages/gasket-plugin-docs/test/utils/generate-index.test.js
@@ -75,8 +75,10 @@ const fullDocsConfigSet = {
 
 const writeFileStub = sinon.stub();
 const generateIndex = proxyquire('../../lib/utils/generate-index', {
-  util: {
-    promisify: () => writeFileStub
+  fs: {
+    promises: {
+      writeFile: writeFileStub
+    }
   }
 });
 const { generateContent } = generateIndex;