+++ packages/gasket-assets/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `@gasket/assets`
 
+### 6.11.2
+
+- Use fs.promises and upgrade mkdirp ([#319])
+
 ### 6.0.0
 
 - Version alignment
@@ -21,3 +25,4 @@
 
 [#120]: https://github.com/godaddy/gasket/pull/120
 [#247]: https://github.com/godaddy/gasket/pull/247
+[#319]: https://github.com/godaddy/gasket/pull/319