+++ packages/gasket-cli/CHANGELOG.md
@@ -1,5 +1,9 @@
 # `@gasket/cli`
 
+### 6.11.2
+
+- Use fs.promises and upgrade mkdirp ([#319])
+
 ### 6.10.0
 
 - TypeScript definitions for config & lifecycles ([#311])
@@ -317,3 +321,4 @@ upon `gasket create`
 [#170]: https://github.com/godaddy/gasket/pull/170
 [#247]: https://github.com/godaddy/gasket/pull/247
 [#311]: https://github.com/godaddy/gasket/pull/311
+[#319]: https://github.com/godaddy/gasket/pull/319